bug修改，但由于so包的缘故，硬解码画面无法播放，重新编译ffmepg的so库包就好。

+++ app/src/main/cpp/PlayerShader.cpp
@@ -189,7 +189,7 @@ bool PlayerShader::init(PlayerShaderType type) {
             1.0f, 1.0f, 0.0f,
             -1.0f, 1.0f, 0.0f,
     };
-    GLuint apos = glGetAttribLocation(program, "aPosition");
+    GLuint apos = (GLuint)glGetAttribLocation(program, "aPosition");
     glEnableVertexAttribArray(apos);
     //传递顶点
     glVertexAttribPointer(apos, 3, GL_FLOAT, GL_FALSE, 12, vers);
